Add wasip1 build support for WASM execution model
Extract processRelocation into build-tagged variants (reloc.go / reloc_wasip1.go)
so the WASM path can use HostMemoryAddress for host-space relocation values. Add
freeShadowAllocations (cleanup.go / cleanup_wasip1.go) to release shadow memory
via VirtualFree after BOF execution. Rename _windows.go files and add
"windows || wasip1" build tags so the package compiles for both targets. Native
build behavior is unchanged.
